Circle Snake
This is a Circle Snake game. You need to control the snake's move to eat the apples on canvas by pressing the SPACE. The snake can move in circle automatically. You can change the rotating direction by pressing the SPACE. Every time the snake eats an apple, it will grow longer. If it go out of the border, it will appear from the opposite side of the canvas. If the snake eats it end, game over.